Add DeepPReLUNet back to components
