+++ src/glass_box_umap/components.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+import math
 
 import torch
 from torch import nn
@@ -22,3 +23,35 @@ class LayerNormDetached(nn.Module):
 
         norm_x = (x - mean) / torch.sqrt(var + 1e-12)  # Added epsilon for stability
         return self.scale * norm_x
+
+
+class DeepPReLUNet(nn.Module):
+    """A network with PReLU activation and LayerNormDetached."""
+
+    def __init__(
+        self,
+        input_dims: tuple[int, ...],
+        n_components: int = 2,
+        hidden_size: int = 256,
+        n_hidden_layers: int = 5,
+    ):
+        super().__init__()
+
+        input_size = math.prod(input_dims)
+        self.flatten = nn.Flatten()
+
+        layers = []
+        for i in range(n_hidden_layers):
+            in_dim = input_size if i == 0 else hidden_size
+
+            layers.append(nn.Linear(in_dim, hidden_size, bias=False))
+            layers.append(nn.PReLU())
+
+            if i < n_hidden_layers - 1:
+                layers.append(LayerNormDetached(hidden_size))
+
+        layers.append(nn.Linear(hidden_size, n_components, bias=False))
+        self.model = nn.Sequential(*layers)
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        return self.model(self.flatten(x))
